requirements.txt
django==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0